local/app/requirements.txt
pywebview
flask
requests
flask-cors